[sailfish-components-weather] Don't depend on application. JB#54526
This was added because weather indicator was left on events view
after uninstalling weather app, but caused a side-effect of being
unable to uninstall the app from app grid due to cyclical dependency.

Better handle the leftover indicator separately. E.g. detect whether
the app itself is there. And could also make sense to merge the
components and the app, but let's get into that a bit later.